requirements.txt
discord.py
asyncio
requests
python-dotenv